requirements/run.txt
ruamel.yaml
docutils=0.17.1
python-dateutil
jinja2
cerberus
xonsh
rever
openpyxl
nameparser
pandas
pymongo
matplotlib-base
habanero
gooey
google-api-python-client
google-auth-oauthlib
